fix: address v0.7 candidate review findings
