+++ cli/trace_loader.go
@@ -1,0 +1,190 @@
+package main
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"io"
+	"net/http"
+	"sort"
+	"strings"
+)
+
+func loadLatestTrace(paths runtimePaths, entityID, domain string) (traceLatestOutput, error) {
+	listOut, err := loadTraceList(paths, entityID, domain)
+	if err != nil {
+		return traceLatestOutput{}, err
+	}
+	if len(listOut.Traces) == 0 {
+		return traceLatestOutput{}, fmt.Errorf("no traces found for %s; Home Assistant keeps only recent traces, and YAML automations/scripts need an id to be traceable", entityID)
+	}
+	latest := listOut.Traces[0]
+	getOut, err := loadTraceGetWithUniqueID(paths, entityID, domain, listOut.UniqueID, latest.RunID)
+	if err != nil {
+		return traceLatestOutput{}, err
+	}
+	return traceLatestOutput{
+		SchemaVersion: 1,
+		OK:            true,
+		EntityID:      entityID,
+		Domain:        domain,
+		UniqueID:      listOut.UniqueID,
+		RunID:         latest.RunID,
+		Timestamp:     latest.Timestamp,
+		LastStep:      getOut.LastStep,
+		Error:         getOut.Error,
+		Trace:         getOut.Trace,
+	}, nil
+}
+
+func loadTraceList(paths runtimePaths, entityID, domain string) (traceListOutput, error) {
+	uniqueID, err := resolveTraceUniqueID(paths, entityID)
+	if err != nil {
+		return traceListOutput{}, err
+	}
+	listBody, err := relayWSJSON(paths, map[string]string{
+		"type":    "trace/list",
+		"domain":  domain,
+		"item_id": uniqueID,
+	})
+	if err != nil {
+		return traceListOutput{}, err
+	}
+	entries, err := parseTraceList(listBody)
+	if err != nil {
+		return traceListOutput{}, err
+	}
+	sort.SliceStable(entries, func(i, j int) bool {
+		return traceTimestampAfter(entries[i].Timestamp, entries[j].Timestamp)
+	})
+	return traceListOutput{
+		SchemaVersion: 1,
+		OK:            true,
+		EntityID:      entityID,
+		Domain:        domain,
+		UniqueID:      uniqueID,
+		Count:         len(entries),
+		Traces:        entries,
+	}, nil
+}
+
+func loadTraceGet(paths runtimePaths, entityID, domain, runID string) (traceGetOutput, error) {
+	uniqueID, err := resolveTraceUniqueID(paths, entityID)
+	if err != nil {
+		return traceGetOutput{}, err
+	}
+	return loadTraceGetWithUniqueID(paths, entityID, domain, uniqueID, runID)
+}
+
+func loadTraceGetWithUniqueID(paths runtimePaths, entityID, domain, uniqueID, runID string) (traceGetOutput, error) {
+	if strings.TrimSpace(runID) == "" {
+		return traceGetOutput{}, fmt.Errorf("trace get requires a run_id from trace list")
+	}
+	traceBody, err := relayWSJSON(paths, map[string]string{
+		"type":    "trace/get",
+		"domain":  domain,
+		"item_id": uniqueID,
+		"run_id":  runID,
+	})
+	if err != nil {
+		return traceGetOutput{}, err
+	}
+	summary := traceGetSummary(traceBody)
+	return traceGetOutput{
+		SchemaVersion:   1,
+		OK:              true,
+		EntityID:        entityID,
+		Domain:          domain,
+		UniqueID:        uniqueID,
+		RunID:           runID,
+		ItemID:          summary.ItemID,
+		Timestamp:       summary.Timestamp,
+		LastStep:        summary.LastStep,
+		State:           summary.State,
+		ScriptExecution: summary.ScriptExecution,
+		Error:           summary.Error,
+		Trace:           extractRelayData(traceBody),
+	}, nil
+}
+
+func resolveTraceUniqueID(paths runtimePaths, entityID string) (string, error) {
+	body, err := relayWSJSON(paths, map[string]string{
+		"type":      "config/entity_registry/get",
+		"entity_id": entityID,
+	})
+	if err != nil {
+		return "", err
+	}
+	var envelope struct {
+		OK   bool `json:"ok"`
+		Data struct {
+			UniqueID string `json:"unique_id"`
+		} `json:"data"`
+		Error struct {
+			Message string `json:"message"`
+		} `json:"error"`
+	}
+	if err := json.Unmarshal(body, &envelope); err != nil {
+		return "", fmt.Errorf("cannot parse entity registry response: %w", err)
+	}
+	if !envelope.OK {
+		if envelope.Error.Message != "" {
+			return "", fmt.Errorf("cannot resolve %s: %s", entityID, envelope.Error.Message)
+		}
+		return "", fmt.Errorf("cannot resolve %s", entityID)
+	}
+	if strings.TrimSpace(envelope.Data.UniqueID) == "" {
+		return "", fmt.Errorf("cannot resolve %s: entity registry entry has no unique_id", entityID)
+	}
+	return envelope.Data.UniqueID, nil
+}
+
+func relayWSJSON(paths runtimePaths, payload any) ([]byte, error) {
+	cfg, err := loadConfig(paths)
+	if err != nil {
+		return nil, err
+	}
+	token, err := readRelayAuthToken()
+	if err != nil {
+		return nil, fmt.Errorf("%s", relayAuthTokenProblemMessage(err))
+	}
+	payloadBytes, err := json.Marshal(payload)
+	if err != nil {
+		return nil, err
+	}
+	url := strings.TrimRight(cfg.RelayBaseURL, "/") + "/ws"
+	req, err := http.NewRequest("POST", url, bytes.NewReader(payloadBytes))
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Set("Authorization", "Bearer "+token)
+	req.Header.Set("Content-Type", "application/json")
+	resp, err := httpClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	if resp.StatusCode >= 400 {
+		return nil, fmt.Errorf("relay ws failed with HTTP %d", resp.StatusCode)
+	}
+	var envelope struct {
+		OK    bool `json:"ok"`
+		Error struct {
+			Message string `json:"message"`
+		} `json:"error"`
+	}
+	if err := json.Unmarshal(body, &envelope); err != nil {
+		return nil, fmt.Errorf("cannot parse relay ws response: %w", err)
+	}
+	if !envelope.OK {
+		if envelope.Error.Message != "" {
+			return nil, fmt.Errorf("relay ws failed: %s", envelope.Error.Message)
+		}
+		return nil, fmt.Errorf("relay ws failed")
+	}
+	return body, nil
+}
